TAL-212 FE | Webapp | Security Configuration
Configured the Angular router so that a non-authenticated user is redirected to the /login page when accessing any page other than /login, /signup or /about. For this a guard was setup in src/app/authenticate-guard.service.ts, which uses a simple mock authentication service located in src/app/authenticate.service.ts. Also added placeholder login and logout buttons to the LoginComponent with which to use the authentication service and to test the redirection when accessing the dashboard page when not logged in.